commit boton borrar y buscar Gusty ok
