+++ src/main/java/com/egg/AppRECC/controladores/PortalControlador.java
@@ -8,7 +8,9 @@ package com.egg.AppRECC.controladores;
 import com.egg.AppRECC.entidades.Posteo;
 import com.egg.AppRECC.excepciones.MiException;
 import com.egg.AppRECC.servicios.PosteoServicio;
+import com.egg.AppRECC.servicios.UsuarioServicio;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -26,67 +28,71 @@ import org.springframework.web.multipart.MultipartFile;
 @Controller
 @RequestMapping("/")
 public class PortalControlador {
-    
+
     @Autowired
     private PosteoServicio posteoServicio;
-    
+
+    @Autowired
+    private UsuarioServicio usuarioServicio;
+
     @GetMapping("/")
     public String index(ModelMap modelo) { //localhost:8080/
-        
+
         //List<Posteo> posteos = posteoServicio.listarPosteos();
         List<Posteo> posteos = posteoServicio.listarPosteosBorrados();
         modelo.addAttribute("posteos", posteos);
 
-        return "nuevo/inicio.html";
+        return "inicio.html";
     }
-    
+
     @GetMapping("nosotros")
     public String nosotros(ModelMap modelo) { //localhost:8080/
-        
+
         //List<Posteo> posteos = posteoServicio.listarPosteos();
         List<Posteo> posteos = posteoServicio.listarPosteosBorrados();
         modelo.addAttribute("posteos", posteos);
 
-        return "nuevo/nosotros.html";
+        return "nosotros.html";
     }
-    
+
     @GetMapping("perfil")
     public String perfil(ModelMap modelo) { //localhost:8080/
-        
+
         //List<Posteo> posteos = posteoServicio.listarPosteos();
         List<Posteo> posteos = posteoServicio.listarPosteosBorrados();
         modelo.addAttribute("posteos", posteos);
 
-        return "nuevo/perfil.html";
+        return "perfil.html";
     }
-    
+
     @GetMapping("publicar")
     public String publicar(ModelMap modelo) { //localhost:8080/
-        
+
         //List<Posteo> posteos = posteoServicio.listarPosteos();
         List<Posteo> posteos = posteoServicio.listarPosteosBorrados();
         modelo.addAttribute("posteos", posteos);
 
-        return "nuevo/publicar.html";
+        return "publicar.html";
     }
+
     @GetMapping("/login")
     public String login(ModelMap modelo) { //localhost:8080/
-        
+
         List<Posteo> posteos = posteoServicio.listarPosteos();
         modelo.addAttribute("posteos", posteos);
 
         return "login.html";
     }
-    
+
     @GetMapping("/register")
     public String register(ModelMap modelo) { //localhost:8080/
-        
+
         List<Posteo> posteos = posteoServicio.listarPosteos();
         modelo.addAttribute("posteos", posteos);
 
         return "register.html";
     }
-    
+
     @PostMapping("/")
     public String Index(@RequestParam("titulo") String titulo,
             @RequestParam("cuerpo") String cuerpo, @RequestParam("file") MultipartFile imagen, ModelMap modelo) {
@@ -96,17 +102,16 @@ public class PortalControlador {
             modelo.put("exito", "la actividad se cargo correctamente");
 
             List<Posteo> posteos = posteoServicio.listarPosteos();
-            
+
             modelo.addAttribute("posteos", posteos);
-            
-            
+
             return "redirect:/";
         } catch (MiException ex) {
             modelo.put("error", ex.getMessage());
-            return "/nuevo/publicar.html";
+            return "publicar.html";
         }
     }
-    
+
     @PostMapping("/modificado")
     public String modificado(@RequestParam("id") Long id, @RequestParam("titulo") String titulo,
             @RequestParam("cuerpo") String cuerpo, @RequestParam("file") MultipartFile imagen, LocalDate fecha, ModelMap modelo) {
@@ -116,7 +121,35 @@ public class PortalControlador {
         modelo.addAttribute("posteos", posteos);
         return "redirect:/";
     }
-    
 
+    @PostMapping("/registro")
+    public String registro(
+            @RequestParam("nombre") String nombre,
+            @RequestParam("email") String email,
+            @RequestParam("password") String password,
+            @RequestParam("password2") String password2,
+            ModelMap modelo
+    ) {
+        try {
+            usuarioServicio.crearUsuario(nombre, email, password, password2);
+            modelo.put("exito", "El usuario se creo correctamente");
+            return "redirect:/";
+        } catch (MiException e) {
+            modelo.put("error", e.getMessage());
+
+            return "register";
+        }
+    }
+
+    @GetMapping("/registrar")
+    public String registrar() {
+        return "register.html";
+    }
     
+    @PostMapping("/buscar")
+    public String buscar(@RequestParam("frase") String frase, ModelMap modelo ) {
+            List<Posteo> posteos = posteoServicio.buscarPosteos(frase);
+            modelo.addAttribute("posteos", posteos);
+        return "buscarPosteos.html";
+    }
 }
