+++ src/main/java/com/egg/AppRECC/repositorios/PosteoRepositorio.java
@@ -21,7 +21,7 @@ public interface PosteoRepositorio extends JpaRepository<Posteo, Long>{
     @Query("SELECT n FROM Posteo n WHERE n.id = :id")
     public Posteo buscarPorId(@Param("id") Long id);
     
-    @Query("SELECT n FROM Posteo n WHERE n.borrado!=true and n.cuerpo LIKE %:frase% ORDER BY n.fecha desc")
+    @Query("SELECT n FROM Posteo n WHERE n.borrado!=true and n.cuerpo LIKE %:frase% and n.titulo LIKE %:frase% ORDER BY n.fecha desc")
     public List<Posteo> buscarPosteos(@Param("frase") String frase);
 
 }
