+++ src/main/resources/templates/login.html
@@ -35,8 +35,8 @@ and open the template in the editor.
                 </div>
             
                 <button>Log In</button>
-                
-                <a th:href="@{/usuario/registrar}">Desea registrarse?</a>
+                <a class="btn btn-secondary" th:href="@{/}">Volver</a>
+                <a th:href="@{/registrar}" class="btn btn-secondary">Desea registrarse?</a>
             </form>
             <p align="center" class="mt-5 mb-3 text-muted">&copy;2022</p>
         </main>
