+++ src/main/resources/templates/publicar.html
@@ -17,6 +17,19 @@
                             </p>
                             <form id="contactform" enctype="multipart/form-data"
                                   th:action="@{/}" method="POST">
+                                <div th:if="${error != null}" class="card text-white bg-danger mb-3 mensajeError">
+                                    <div class="card-body">
+                                        <h5 class="card-title">Error</h5>
+                                        <p class="card-text" th:text="${error}"></p>
+                                    </div>    
+                                </div>
+
+                                <div th:if="${exito != null}" class="card text-white bg-succes mb-3 mt-3 mensajeExito">
+                                    <div class="card-body">
+                                        <h5 class="card-title">Exito</h5>
+                                        <p class="card-text" th:text="${exito}"></p>
+                                    </div>
+                                </div>
                                 <ol>
                                     <li>
                                         <label for="name" th>Titulo</label>
